requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-dotenv==1.0.0
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
requests==2.31.0
beautifulsoup4==4.12.2
selenium==4.15.2
playwright==1.40.0
openai==1.3.7
python-telegram-bot==20.7
pydantic==2.5.0
python-multipart==0.0.6
aiofiles==23.2.1
flask==2.2.5
flask-cors==3.0.10
